+++ packages/database/src/schema/index.ts
@@ -2,6 +2,7 @@ export * from "./sites.js";
 export * from "./users.js";
 export * from "./access-control.js";
 export * from "./content.js";
+export * from "./content-schedule-events.js";
 export * from "./content-types.js";
 export * from "./revisions.js";
 export * from "./settings.js";